agents-adk/adk-bidi-streaming-/requirements.txt
google-adk>=0.2.0
fastapi
uvicorn[standard]
python-dotenv
certifi
